requirements.txt
cloudpickle==1.6.0
googleapis-common-protos==1.51.0
grpcio==1.26.0
grpcio-tools==1.26.0
linselect==0.1.1
matplotlib==3.4.2
numpy==1.20.3
pandas==1.2.4
pyarrow==4.0.0
pyperf==2.2.0
pytest==6.2.4
saxpy==1.0.1.dev167
scikit-learn==0.20.2
scipy==1.6.3
setuptools==57.0.0
skidmarks==0.0.6
statsmodels==0.12.2
stumpy==1.8.0
tqdm==4.59
tox==3.23.0
tox-pyenv==1.1.0
typing-extensions==3.10.0.0
debugpy